backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi-pagination>=0.15.4",
    "fastapi[standard]>=0.122.0",
    "faststream>=0.6.4",
    "latex2mathml>=3.78.1",
    "markdown-it-py[plugins]>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "pygments>=2.17.0",
    "orjson>=3.11.4",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.4",
    "scalar-fastapi>=1.5.0",
    "sqlmodel>=0.0.27",
    "uuid6>=2025.0.1",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "ipdb>=0.13.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
monitoring = [
    "sentry-sdk[fastapi]>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[tool.pytest.ini_options]
addopts = "-p no:warnings"
asyncio_mode = "auto"
# 关键配置：所有异步测试和 fixtures 共用同一个事件循环
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "unit: mark test as a unit test",
    "integration: mark test as an integration test",
    "middleware: mark test as a middleware test",
    "users: mark test as a users module test",
    "permissions: mark test as a permissions related test",
    "media: mark test as a media module test",
    "posts: mark test as a posts module test",
    "git_ops: mark test as a git_ops module test",

]

# ==========================================
# 测试覆盖率配置
# ==========================================
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/conftest.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "app/initial_data.py",  # 初始化脚本
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "@abstractmethod",
    "@overload",
    "if settings.DEBUG:",
]

# 设置覆盖率阈值（低于此值测试失败）
fail_under = 70

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.1.1",
    "httpx>=0.28.1",
    "rich>=14.2.0",
    "pytest-mock>=3.15.1",
]

[tool.setuptools]
packages = ["app"]
